+++ src/main/java/life/majiang/community/service/QuestionService.java
@@ -1,6 +1,5 @@
 package life.majiang.community.service;
 
-import com.sun.tools.javac.comp.Annotate;
 import life.majiang.community.dto.PageDTO;
 import life.majiang.community.dto.QuestionDTO;
 import life.majiang.community.exception.CustomizeErrorCode;import life.majiang.community.exception.CustomizeException;
@@ -10,7 +9,6 @@ import life.majiang.community.mapper.UserMapper;
 import life.majiang.community.model.Question;
 import life.majiang.community.model.QuestionExample;
 import life.majiang.community.model.User;
-import life.majiang.community.model.UserExample;
 import org.apache.ibatis.session.RowBounds;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -66,7 +64,7 @@ public class QuestionService {
 
         return pageDTO;
     }
-    public PageDTO list(Integer id,Integer page,Integer size){
+    public PageDTO list(Long id, Integer page, Integer size){
         PageDTO pageDTO=new PageDTO();
         Integer totalPage;
         QuestionExample example=new QuestionExample();
@@ -101,7 +99,7 @@ public class QuestionService {
         return pageDTO;
     }
 
-    public QuestionDTO getById(Integer id) {
+    public QuestionDTO getById(Long id) {
        Question question= questionMapper.selectByPrimaryKey(id);
        if (question == null){
            throw  new CustomizeException(CustomizeErrorCode.QUESTION_NOT_FOUND);
@@ -137,7 +135,7 @@ public class QuestionService {
 
     }
 
-    public void incView(Integer id) {
+    public void incView(Long id) {
         Question record = new Question();
         record.setViewCount(1);
         record.setId(id);
